Remove unused Open Graph image generation code
